김홍준 VO,DTO, 경로 Config파일에 수정완료- 정상동작확인
