+++ parkingmate/src/main/java/com/app/parkingmate/domain/DTO/CarInfoDTO.java
@@ -1,52 +1,39 @@
-package com.app.parkingmate.domain.VO;
+package com.app.parkingmate.domain.DTO;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 
 @Component
 @Data
-public class ParkingDTO {
-//    주차장 고유 id
-//    not null
-    private Integer ParkingId;
-//    주차장 주소
-//    3000자
-//    not null
-    private String ParkingAddress;
-//    주차장 카테고리
-//    1000자
-//    도시(시),(구)를 나타냄
-    private String ParkingCategory;
-//    주차장 명
-//    1000자
-//    not null
-    private String ParkingName;
-//    주차장내용(세부정보)
-//    3000자
-    private String ParkingContent;
-//    주차장 평점
-//    기본 값 : 5
-    private Integer ParkingGrade;
-//    주차장 이용 수
-//    기본 값 : 0
-    private  Integer ParkingUseCount;
-//    주차장 대여 금액(시간당)
-//    기본 값 : 1000
-    private Integer ParkingPrice;
-//    주차 면수
-//    기본 값 : 5
-//    not null
-    private Integer Parking_Count;
+public class CarInfoDTO {
+//    차량정보 고유 id
+//     not null
+    private Integer CarInfoId;
+//    차량 모델
+//     not null
+    private String CarModel;
+//    차량 번호
+//     not null
+    private String CarNumber;
+//    차량 등록일
+//     not null
+    private Date CarRegisterDate;
+//    차량 등록 상태
+//    기본 값 : 등록 중
+//    옵션 : 등록, 취소, 알 수 없음
+//     not null
+    private String CarRegisterStatus;
 //    생성 날짜
-    private Date CreateDate;
+    private LocalDateTime CreateDate;
 //    업데이트 날짜
-    private Date UpdateDate;
+    private LocalDateTime UpdateDate;
 
-//    =================================== FK
-//    회원 테이블 회원 번호
-//    not null
+//    ================================ FK
+//    회원 테이블 회원 id
+//     not null
     private Integer UserId;
     //    회원 닉네임
 //    not null
