+++ parkingmate/src/main/java/com/app/parkingmate/domain/DTO/ParkingDTO.java
@@ -1,34 +1,51 @@
-package com.app.parkingmate.domain.VO;
+package com.app.parkingmate.domain.DTO;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
 
-import java.time.LocalDateTime;
 import java.util.Date;
 
 @Component
 @Data
-public class ProfileDTO {
-//    프로필 고유 id
+public class ParkingDTO {
+//    주차장 고유 id
 //    not null
-    private Integer ProfileId;
-//    프로필 명
-//    not null
-    private String ProfileName;
-//    프로필 내용
-//    기본 값 : 안녕하세요
-//    3000자
-    private String ProfileContent;
-//    프로필 이미지 경로
+    private Integer ParkingId;
+//    주차장 주소
 //    3000자
 //    not null
-    private String ProfileImagePath;
+    private String ParkingAddress;
+//    주차장 카테고리
+//    1000자
+//    도시(시),(구)를 나타냄
+    private String ParkingCategory;
+//    주차장 명
+//    1000자
+//    not null
+    private String ParkingName;
+//    주차장내용(세부정보)
+//    3000자
+    private String ParkingContent;
+//    주차장 평점
+//    기본 값 : 5
+    private Integer ParkingGrade;
+//    주차장 이용 수
+//    기본 값 : 0
+    private  Integer ParkingUseCount;
+//    주차장 대여 금액(시간당)
+//    기본 값 : 1000
+    private Integer ParkingPrice;
+//    주차 면수
+//    기본 값 : 5
+//    not null
+    private Integer Parking_Count;
 //    생성 날짜
-    private LocalDateTime CreateDate;
+    private Date CreateDate;
 //    업데이트 날짜
-    private LocalDateTime UpdateDate;
-//    =========================================== FK
-//    회원 테이블 고유 id
+    private Date UpdateDate;
+
+//    =================================== FK
+//    회원 테이블 회원 번호
 //    not null
     private Integer UserId;
     //    회원 닉네임
