+++ parkingmate/src/main/java/com/app/parkingmate/domain/DTO/ProfileDTO.java
@@ -1,4 +1,4 @@
-package com.app.parkingmate.domain.VO;
+package com.app.parkingmate.domain.DTO;
 
 import lombok.Data;
 import org.springframework.stereotype.Component;
@@ -8,57 +8,27 @@ import java.util.Date;
 
 @Component
 @Data
-public class ParkingPictureDTO {
-//    주차장 사진 고유 번호
+public class ProfileDTO {
+//    프로필 고유 id
 //    not null
-    private Integer ParkingPictureId;
-//    주차장 사진 명
+    private Integer ProfileId;
+//    프로필 명
 //    not null
-//    500자
-    private String ParkingPictureName;
-//    주차장 사진 경로
+    private String ProfileName;
+//    프로필 내용
+//    기본 값 : 안녕하세요
+//    3000자
+    private String ProfileContent;
+//    프로필 이미지 경로
 //    3000자
 //    not null
-    private String ParkingPicturePath;
-//    생성날짜
+    private String ProfileImagePath;
+//    생성 날짜
     private LocalDateTime CreateDate;
 //    업데이트 날짜
     private LocalDateTime UpdateDate;
-
 //    =========================================== FK
-//    주차장 테이블 고유번호
-//    not null
-    private Integer ParkingId;
-    //    주차장 주소
-//    3000자
-//    not null
-    private String ParkingAddress;
-    //    주차장 카테고리
-//    1000자
-//    도시(시),(구)를 나타냄
-    private String ParkingCategory;
-    //    주차장 명
-//    1000자
-//    not null
-    private String ParkingName;
-    //    주차장내용(세부정보)
-//    3000자
-    private String ParkingContent;
-    //    주차장 평점
-//    기본 값 : 5
-    private Integer ParkingGrade;
-    //    주차장 이용 수
-//    기본 값 : 0
-    private  Integer ParkingUseCount;
-    //    주차장 대여 금액(시간당)
-//    기본 값 : 1000
-    private Integer ParkingPrice;
-    //    주차 면수
-//    기본 값 : 5
-//    not null
-    private Integer Parking_Count;
-//==========================================================
-//    회원 테이블 고유 번호
+//    회원 테이블 고유 id
 //    not null
     private Integer UserId;
     //    회원 닉네임
